Squashed some bugs
Fixes an error with the Steam Page not loading the game ID correctly.
(URL format might have changed and I missed it?) Now uses a better
regexp approach, whcih should be better long-term anyway.

Added a new feature of displaying the results in progress. In the event
the scan stalls/halts, you'll still be able to get partial results.